+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 6.8.0 (December 1, 2018)
 ### New Features
+- [#898](https://github.com/openscope/openscope/issues/898) - Add command to exit holding pattern